final_debug_percent
