+++ srcs/flags/flag_width.c
@@ -6,14 +6,14 @@
 /*   By: mbos <mbos@student.le-101.fr>              +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/15 12:02:25 by mbos         #+#   ##    ##    #+#       */
-/*   Updated: 2019/12/18 18:18:30 by mbos        ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/12/18 18:29:26 by mbos        ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "../../includes/ft_printf.h"
 
-t_bool case_percent(t_flag *l_flag, char *width, va_list params)
+t_bool		case_percent(t_flag *l_flag, char *width, va_list params)
 {
 	if (width[1] == '*')
 	{
@@ -25,7 +25,10 @@ t_bool case_percent(t_flag *l_flag, char *width, va_list params)
 			if (l_flag->type != '-')
 				l_flag->type = '-';
 		}
+		return (True);
 	}
+	if (!(l_flag->width = get_width_preci(width)))
+		return (false_ret(__func__));
 	return (True);
 }
 
@@ -43,9 +46,9 @@ t_bool		flag_width(t_flag *l_flag, char *width, va_list params)
 		}
 		return (True);
 	}
-	if ((l_flag->type == '-' && width[0] == '0') 
+	if ((l_flag->type == '-' && width[0] == '0')
 			|| (l_flag->type == '0' && width[0] == '-'))
-		case_percent(l_flag, width, params);
+		return (case_percent(l_flag, width, params));
 	if (!(l_flag->width = get_width_preci(width)))
 		return (false_ret(__func__));
 	return (True);
